Project Company and Structure rework
To allow for a simpler pull from git for non-program-literate people to understand the file structure. The project structure has been simplified to separate the independent assets of the project into individual folders.
Also the package name has been updated to include the company name "ratbox" within the android app