No bluetooth hardware exception try-catch
On android devices that do not have bluetooth hardware, trying to display the bluetooth pop-up dialog will cause a crash.
This has been nested into a try-catch. Should an exception be thrown due to lack of hardware, a simple toast will display to the user for feedback about this.

Additionally should there be a case where the user does not have any devices paired to the android device. A toast will occur instructing the user to first pair their synthvisor with their android device