SendtoVisor Button created
Due to the transfer speed of the bluetooth module and the processing, it was better to add a universal "Upload to visor" button which will process the entire app's config to send to the device. Current Dialog box is a WIP, but currently does send to the device.

There seems to be an issue on appearance of the upload dialog box that the first button press of "upload" will not function:

"W/InputEventReceiver: Attempted to finish an input event but the input input event receiver has already been disposed"