Tab2 RecyclerView functionality implemented
The card recycler gridView for the second tab "HUD" to see all saved patterns now functions and will display all saved patterns on screen @onCreate().

Doing so it has been discovered that each tab is going to have to read from the pattern text file in the future. so the synthPattern static class object has been removed and is initialised, separately, on each page.

Going forward, all object will be reading like this. So the other static methods (with the exception of BluetoothManager) will follow the same process in the future.